requirements_app.txt
torch==2.3.1
numpy==1.26.4
fastapi==0.111.0
joblib==1.4.2
pillow==10.3.0
scikit-learn==1.5.0
